Atualização geral para correções de bugs

+++ src/main/java/com/blogpessoal/Turma29/controller/PostagemController.java
@@ -42,7 +42,7 @@ public class PostagemController {
 	 * Busca pelo ID
 	 */
 	@GetMapping(value = "/{id}")
-	public ResponseEntity<Postagem> findById(@PathVariable Integer id){
+	public ResponseEntity<Postagem> findById(@PathVariable Long id){
 		ResponseEntity<Postagem> obj = service.findById(id);
 		return obj;
 	}
@@ -76,7 +76,7 @@ public class PostagemController {
 	 * Deleta uma postagem
 	 */
 	@DeleteMapping("/{id}")
-	public void Delete(@PathVariable Integer id) {
+	public void Delete(@PathVariable Long id) {
 		service.delete(id);
 	}
