+++ src/main/java/com/blogpessoal/Turma29/repository/PostagemRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.blogpessoal.Turma29.model.Postagem;
 
 @Repository
-public interface PostagemRepository extends JpaRepository<Postagem, Integer>{
+public interface PostagemRepository extends JpaRepository<Postagem, Long>{
 	
 	public List<Postagem> findAllByTituloContainingIgnoreCase (String titulo);
